requirements.txt
# 🚀 Dependencias del Backend - Universal Trading Analyzer

# Framework web principal
fastapi==0.104.1          # Framework web moderno y rápido
uvicorn[standard]==0.24.0 # Servidor ASGI para ejecutar FastAPI

# Validación y configuración
pydantic==2.11.7          # Validación de datos (versión compatible)
python-dotenv==1.0.0      # Manejo de variables de entorno

# Análisis técnico y datos financieros
pandas==2.3.1             # Manipulación de datos financieros (compatible)
numpy>=1.26.0,<2.0        # Cálculos numéricos (compatible con pandas-ta)
pandas-ta==0.3.14b0       # ✅ REEMPLAZO: 150+ indicadores técnicos
# ccxt removed - using Capital.com API directly

# Comunicación en tiempo real
websockets>=15.0          # WebSockets nativos de Python (actualizado)
aiohttp>=3.12.0           # Cliente HTTP asíncrono (actualizado)

# Utilidades
python-json-logger==3.3.0 # Logging estructurado (actualizado)
PyYAML>=6.0.2             # Manejo de archivos de configuración

# Visualización y gráficos para backtesting
matplotlib>=3.8.0        # Gráficos y visualización
seaborn>=0.13.0           # Gráficos estadísticos avanzados
plotly>=5.17.0            # Gráficos interactivos

# Base de datos
sqlalchemy>=2.0.0         # ORM para base de datos

# Testing (para desarrollo)
pytest>=8.4.0            # Framework de testing (actualizado)
pytest-asyncio>=1.1.0    # Testing para código asíncrono (actualizado)
httpx>=0.28.0             # Cliente HTTP para testing (actualizado)

# Display en tiempo real para live trading bot
colorama>=0.4.6           # Colores en terminal
rich>=13.7.0              # Display avanzado en terminal